requirements.txt
# LangChain Core and Framework
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=1.0.0

# Vector Database
chromadb>=0.4.22

# Stock Data
yfinance>=0.2.35

# Web Scraping and News
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.1.0

# Data Processing
pandas>=2.2.0
numpy>=1.26.0,<2.0.0  # Pin to 1.x for PyTorch compatibility

# NLP and Embeddings
sentence-transformers>=2.3.0
nltk>=3.8.0

# Configuration
python-dotenv>=1.0.0

# Utilities
pydantic>=2.6.0
python-dateutil>=2.8.2
